restructure to match main

+++ src/actions/db/users.ts
@@ -1,5 +1,7 @@
 import type { NewUser, Users } from "@/src/types";
-import { supabase } from "@/src/lib/supabase-client";
+import { createServerClient } from "@/src/lib/supabase-client";
+
+const supabase = await createServerClient();
 
 export async function createUser(user: NewUser) {
   const { error } = await supabase.from("users").insert(user);
@@ -10,7 +12,7 @@ export async function createUser(user: NewUser) {
   }
 }
 
-export async function readUser(userId: number): Promise<Users | null> {
+export async function readUser(userId: string): Promise<Users | null> {
   const { data, error } = await supabase
     .from("users")
     .select("*")
@@ -25,7 +27,7 @@ export async function readUser(userId: number): Promise<Users | null> {
   return data;
 }
 
-export async function updateUser(userId: number, updatedUser: Partial<Users>) {
+export async function updateUser(userId: string, updatedUser: Partial<Users>) {
   const { error } = await supabase
     .from("users")
     .update(updatedUser)
@@ -37,11 +39,8 @@ export async function updateUser(userId: number, updatedUser: Partial<Users>) {
   }
 }
 
-export async function deleteUser(userId: number) {
-  const { error } = await supabase
-    .from("users")
-    .delete()
-    .eq("id", userId);
+export async function deleteUser(userId: string) {
+  const { error } = await supabase.from("users").delete().eq("id", userId);
 
   if (error) {
     console.error("Error deleting user:", error.message);
